+++ src/app/AvailablePilets.tsx
@@ -1,22 +1,28 @@
 import { FC } from 'react';
-import { ListGroup, ListGroupItem, Button } from 'reactstrap';
+import { ListGroup, ListGroupItem, Button, CustomInput } from 'reactstrap';
 import { jsx } from '@emotion/core';
-import { removePilet } from './commands';
+import { removePilet, togglePilet } from './commands';
 import { useStore } from './store';
 
 export interface AvailablePiletsProps {}
 
 export const AvailablePilets: FC<AvailablePiletsProps> = () => {
-  const pilets = useStore(m => m.state.pilets);
+  const pilets = useStore(m => m.state.pilets) ?? [];
+
   return (
     <ListGroup>
-      {pilets &&
-        pilets.map(pilet => (
-          <ListGroupItem key={pilet.name}>
-            <Button onClick={() => removePilet(pilet.name)} close />
-            {pilet.name}
-          </ListGroupItem>
-        ))}
+      {pilets.map(pilet => (
+        <ListGroupItem key={pilet.name}>
+          <Button onClick={() => removePilet(pilet.name)} close />
+          <CustomInput
+            type="switch"
+            id={`toggle-${pilet.name}`}
+            label={pilet.name}
+            checked={!pilet.disabled}
+            onChange={() => togglePilet(pilet.name)}
+          />
+        </ListGroupItem>
+      ))}
     </ListGroup>
   );
 };